+++ sdk.class.php
@@ -115,8 +115,8 @@ function __aws_sdk_ua_callback()
 // INTERMEDIARY CONSTANTS
 
 define('CFRUNTIME_NAME', 'aws-sdk-php');
-define('CFRUNTIME_VERSION', '1.5.10');
-define('CFRUNTIME_BUILD', '20120717150000');
+define('CFRUNTIME_VERSION', '1.5.11');
+define('CFRUNTIME_BUILD', '20120801150000');
 define('CFRUNTIME_USERAGENT', CFRUNTIME_NAME . '/' . CFRUNTIME_VERSION . ' PHP/' . PHP_VERSION . ' ' . str_replace(' ', '_', php_uname('s')) . '/' . str_replace(' ', '_', php_uname('r')) . ' Arch/' . php_uname('m') . ' SAPI/' . php_sapi_name() . ' Integer/' . PHP_INT_MAX . ' Build/' . CFRUNTIME_BUILD . __aws_sdk_ua_callback());
 
 
@@ -1439,10 +1439,10 @@ class CFLoader
 			return false;
 		}
 
-		// Load array-to-domdocument
-		elseif (strstr($class, 'Array2DOM'))
+		// Load Transmogrifier
+		elseif (strstr($class, 'Transmogrifier'))
 		{
-			if (file_exists($require_this = $path . 'lib' . DIRECTORY_SEPARATOR . 'dom' . DIRECTORY_SEPARATOR . 'ArrayToDOMDocument.php'))
+			if (file_exists($require_this = $path . 'lib' . DIRECTORY_SEPARATOR . 'dom' . DIRECTORY_SEPARATOR . 'Transmogrifier.php'))
 			{
 				require_once $require_this;
 				return true;